% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandatafunction.R
\name{CleanPetData}
\alias{CleanPetData}
\title{Clean Pet Data}
\usage{
CleanPetData(example_data)
}
\arguments{
\item{example_data}{pet registration data to clean, should include columns of ZIP.Code, Species, Primary.Breed}
}
\description{
Cleans Pet Registration Data
}
\examples{
CleanPetData(example_data)
}
\keyword{clean,}
\keyword{pet}
