% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petpiechart.R
\name{PetPieChart}
\alias{PetPieChart}
\title{Pet Pie Chart}
\usage{
PetPieChart(pet_licenses, zipcode, species_name)
}
\arguments{
\item{pet_licenses}{cleaned dataframe of pet registration data with columns of ZIPCode, Species, Primary.Breed}

\item{zipcode}{the ZIP code of interest, should be in format "00000"}

\item{species_name}{"Dog" or "Cat", depending on which pet species want to examine}
}
\description{
Provides pie chart breakdown of breeds of dogs or cats in specific ZIP code
}
\examples{
PetPieChart(pet_licenses,"98004","Dog")
}
\keyword{breakdown,}
\keyword{chart,}
\keyword{pet}
\keyword{pie,}
