% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortzipfunction.R
\name{SortZIPCodes}
\alias{SortZIPCodes}
\title{Sort ZIP Codes}
\usage{
SortZIPCodes(pet_licenses)
}
\arguments{
\item{pet_licenses}{cleaned dataframe of pet registration data with columns of ZIPCode, Species, Primary.Breed}
}
\description{
Sorts pet registration data by ZIP code, provides total number of registered pets and dog:cat ratio by ZIP code
}
\examples{
SortZIPCodes(pet_licenses)
}
\keyword{pet}
\keyword{ratio,}
\keyword{sort,}
